Move integration tests to their own directory
