+++ proving_system/stark/src/air/context.rs
@@ -26,7 +26,7 @@ impl AirContext {
 
 #[derive(Clone, Debug)]
 pub struct ProofOptions {
-    pub(crate) blowup_factor: u8,
+    pub blowup_factor: u8,
     pub fri_number_of_queries: u8,
     pub coset_offset: u64,
 }
